remove dataset iiif exceptions
