+++ spec/presenters/document/web_resource_presenter_spec.rb
@@ -84,29 +84,67 @@ RSpec.describe Document::WebResourcePresenter do
 
   context 'record with iiif webresource' do
     let(:iif_service) { 'http://iiif.example.org/iiif' }
-    let(:api_response) do
-      base_api_response.tap do |response|
-        response[:object][:services] = [
-          {
-            'about': iif_service,
-            'dctermsConformsTo': [
-              'http://iiif.io/api/image'
-            ]
-          }
-        ]
-        response[:object][:aggregations].first.tap do |agg|
-          agg[:edmIsShownBy] = wr_url
-          agg[:webResources].first.tap do |wr|
-            wr[:svcsHasService] = [iif_service]
-            wr[:ebucoreHasMimeType] = 'image/jpeg'
+    context 'when it is a single page' do
+      let(:api_response) do
+        base_api_response.tap do |response|
+          response[:object][:services] = [
+            {
+              'about': iif_service,
+              'dctermsConformsTo': [
+                'http://iiif.io/api/image'
+              ]
+            }
+          ]
+          response[:object][:aggregations].first.tap do |agg|
+            agg[:edmIsShownBy] = wr_url
+            agg[:webResources].first.tap do |wr|
+              wr[:svcsHasService] = [iif_service]
+              wr[:ebucoreHasMimeType] = 'image/jpeg'
+            end
           end
         end
       end
+
+      it { is_expected.to be_iiif }
+      it { is_expected.to be_displayable }
+      it { is_expected.to be_playable }
+      it { is_expected.to be_downloadable }
+
+      it 'should have the manifest' do
+        expect(subject.iiif_manifest).to eq('http://iiif.example.org/iiif/info.json')
+      end
     end
 
-    it { is_expected.to be_iiif }
-    it { is_expected.to be_displayable }
-    it { is_expected.to be_playable }
-    it { is_expected.to be_downloadable }
+    context 'when it has a manifest' do
+      let(:api_response) do
+        base_api_response.tap do |response|
+          response[:object][:services] = [
+            {
+              'about': iif_service,
+              'dctermsConformsTo': [
+                'http://iiif.io/api/image'
+              ]
+            }
+          ]
+          response[:object][:aggregations].first.tap do |agg|
+            agg[:edmIsShownBy] = wr_url
+            agg[:webResources].first.tap do |wr|
+              wr[:svcsHasService] = [iif_service]
+              wr[:dctermsIsReferencedBy] = 'https://www.example.org/iiif/manifest.json'
+              wr[:ebucoreHasMimeType] = 'image/jpeg'
+            end
+          end
+        end
+      end
+
+      it { is_expected.to be_iiif }
+      it { is_expected.to be_displayable }
+      it { is_expected.to be_playable }
+      it { is_expected.to be_downloadable }
+
+      it 'should have the manifest' do
+        expect(subject.iiif_manifest).to eq('https://www.example.org/iiif/manifest.json')
+      end
+    end
   end
 end
